move get_*_info.py to lake/; create new dir subgraph/ and move all *subgraph*.py into it
