+++ pdr_backend/analytics/get_predictoors_info.py
@@ -5,8 +5,8 @@ from enforce_typing import enforce_types
 from pdr_backend.ppss.ppss import PPSS
 from pdr_backend.util.csvs import save_prediction_csv
 from pdr_backend.util.networkutil import get_sapphire_postfix
-from pdr_backend.util.predictoor_stats import get_cli_statistics
-from pdr_backend.util.subgraph_predictions import (
+from pdr_backend.analytics.predictoor_stats import get_cli_statistics
+from pdr_backend.subgraph.subgraph_predictions import (
     fetch_filtered_predictions,
     FilterMode,
 )
