+++ pdr_backend/analytics/test/test_get_predictoors_info.py
@@ -4,8 +4,8 @@ from enforce_typing import enforce_types
 
 from pdr_backend.ppss.ppss import mock_ppss
 from pdr_backend.ppss.web3_pp import del_network_override
-from pdr_backend.util.get_predictoors_info import get_predictoors_info_main
-from pdr_backend.util.subgraph_predictions import FilterMode
+from pdr_backend.analytics.get_predictoors_info import get_predictoors_info_main
+from pdr_backend.subgraph.subgraph_predictions import FilterMode
 
 
 @enforce_types
@@ -17,7 +17,7 @@ def test_get_predictoors_info_main_mainnet(tmpdir, monkeypatch):
     mock_save = Mock()
     mock_getstats = Mock()
 
-    PATH = "pdr_backend.util.get_predictoors_info"
+    PATH = "pdr_backend.analytics.get_predictoors_info"
     with patch(f"{PATH}.fetch_filtered_predictions", mock_fetch), patch(
         f"{PATH}.save_prediction_csv", mock_save
     ), patch(f"{PATH}.get_cli_statistics", mock_getstats):
