+++ pdr_backend/lake/gql_data_factory.py
@@ -14,7 +14,7 @@ from pdr_backend.lake.table_pdr_predictions import (
 )
 from pdr_backend.ppss.ppss import PPSS
 from pdr_backend.util.networkutil import get_sapphire_postfix
-from pdr_backend.util.subgraph_predictions import get_all_contract_ids_by_owner
+from pdr_backend.subgraph.subgraph_predictions import get_all_contract_ids_by_owner
 from pdr_backend.util.timeutil import pretty_timestr, current_ut
