+++ pdr_backend/lake/test/test_gql_data_factory.py
@@ -7,7 +7,7 @@ import polars as pl
 from pdr_backend.lake.test.resources import _gql_data_factory
 from pdr_backend.lake.table_pdr_predictions import predictions_schema
 from pdr_backend.ppss.web3_pp import del_network_override
-from pdr_backend.util.subgraph_predictions import FilterMode
+from pdr_backend.subgraph.subgraph_predictions import FilterMode
 from pdr_backend.util.timeutil import timestr_to_ut
 
 # ====================================================================
