+++ pdr_backend/payout/payout.py
@@ -7,7 +7,10 @@ from pdr_backend.models.predictoor_contract import PredictoorContract
 from pdr_backend.models.wrapped_token import WrappedToken
 from pdr_backend.ppss.ppss import PPSS
 from pdr_backend.util.constants import SAPPHIRE_MAINNET_CHAINID
-from pdr_backend.util.subgraph import query_pending_payouts, wait_until_subgraph_syncs
+from pdr_backend.subgraph.core_subgraph import (
+    query_pending_payouts,
+    wait_until_subgraph_syncs,
+)
 
 
 @enforce_types
