+++ pdr_backend/util/check_network.py
@@ -6,7 +6,10 @@ from enforce_typing import enforce_types
 from pdr_backend.models.token import Token
 from pdr_backend.ppss.ppss import PPSS
 from pdr_backend.util.constants_opf_addrs import get_opf_addresses
-from pdr_backend.util.subgraph import get_consume_so_far_per_contract, query_subgraph
+from pdr_backend.subgraph.core_subgraph import (
+    get_consume_so_far_per_contract,
+    query_subgraph,
+)
 
 WEEK = 86400 * 7
